requirements.txt
Flask==3.1.2
Flask-Limiter==3.5.0
Flask-WTF==1.2.2
python-dotenv==1.2.1
WTForms==3.2.1
Werkzeug==3.1.3
Jinja2==3.1.6
MarkupSafe==3.0.3
itsdangerous==2.2.0
click==8.3.1
Flask-Login
gunicorn
psycopg2-binary
